Use EventCache + built in mobly SnippetEvents to wait for SMS sent confirmation.
This shares code with EventSnippet in Mobly Snippet Lib. I'm not sure what the accepted practice is for sharing code across github projects? Should I refactor some of the event code in Mobly Bundle Lib into EventUtils, then use that in Mobly Bundled Snippets to avoid duplicating this?

I still want to check this is in as is, to have a working version earlier.