Include IdentityAddress for BluetoothDevice in JsonSerializer.java
When Bt bonds over LE the, getAddress() on btGetPairedDevices() will return a randomized bt address. This causes asserts in bt bonding because the device is comparing the randomized public address with its private identity address. Adding the identity address in the bluetooth device bundle will allow validation that devices are bonded